move mocks to generators namespace
