+++ test/electionguard/generators/ballot.cpp
@@ -16,7 +16,8 @@ EG_API eg_electionguard_status_t
 eg_test_ballot_mocks_get_simple_ballot_from_file(eg_plaintext_ballot_t **out_handle)
 {
     try {
-        auto election = electionguard::test::mocks::BallotGenerator::getSimpleBallotFromFile();
+        auto election =
+          electionguard::tools::generators::BallotGenerator::getSimpleBallotFromFile();
         *out_handle = AS_TYPE(eg_plaintext_ballot_t, election.release());
         return ELECTIONGUARD_STATUS_SUCCESS;
     } catch (const exception &e) {
