+++ test/electionguard/test_manifest.c
@@ -1,5 +1,5 @@
-#include "mocks/election.h"
-#include "mocks/manifest.h"
+#include "generators/election.h"
+#include "generators/manifest.h"
 #include "utils/utils.h"
 
 #include <assert.h>
